Make page level an unsigned value.
A negative level isn't valid and expressing that in the type allows the
analyzers to better understand the code.

+++ src/arch/aarch64/inc/hf/arch/mm.h
@@ -202,7 +202,7 @@ static inline void arch_mm_set_vm(uint64_t vmid, paddr_t table)
 
 uint64_t arch_mm_mode_to_attrs(int mode);
 bool arch_mm_init(paddr_t table, bool first);
-int arch_mm_max_level(int mode);
+uint8_t arch_mm_max_level(int mode);
 
 /**
  * Given the attrs from a table at some level and the attrs from all the blocks
